Add a note for the Platforms folder
The files for the .csproj did not work within the Platforms folder. I lost a few weeks sulking over that and trying to find a workaround. Hopefully with that note others will place their files somewhere else from the start.

Preview does not work for the [!NOTE]. I want it to look as the [!IMPORTANT]. I just copied that and the preview for [!IMPORTANT] also doesn't work, so I guess we're good.