Add caution about iOS AppIcon background color
Currently, App Store Connects will systematically fails when a `<MauiIcon>` is added without any `Color` value. By default, the background is then considered as transparent, which triggers a 100% fails while uploading the archive to App Store Connect (even by using Xcode or command line tools provided by Apple).
In order to fix that, related to this [issue](https://github.com/dotnet/maui/issues/11124) in the `maui` repository, it can be very helpful to warn users that a background must be provided. I tested on `svg` and `png` format, this solved the issue as mentioned in the linked issue.

For this reason, applying a **CAUTION** is, in my opinion, relevant. We are saving lot's of time to developers and avoiding tremendous frustration.