Fixed .NET Upgrade Assistant installation instruction (#1278)
There was a copy/paste error in the installation section for .NET Upgrade Assistant. The installation instructions said that the tool should be installed with the following command:

`dotnet tool update -g upgrade-assistant`

In order to install the tool you have to use `install` instead of `update`. So the correct command is:

`dotnet tool install -g upgrade-assistant`