Update NotePage.xaml.cs
Bing AI told me that this is a syntax error, not a shorthand feature. In C#, if you use an if statement, you must use curly braces {} to indicate the code block that follows, even if there is only one line of code, you cannot omit the curly braces {}. Otherwise, beginners who follow the example documentation to modify the code can easily cause logic errors or variable scope problems.